glossary-service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.3
grpcio-health-checking==1.60.0